// src/app/core/models/masters/audit-management/am-audit-report-type.ts
export interface AmAuditReportType {
    id: number;
    title: string;
    type: string;
    status: string;
    status_id: number;
    status_label: string;
}

export interface AmAuditReportTypePaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: AmAuditReportType[];
}